requirements.txt
fpdf2>=2.7.0
Pillow>=10.0.0
